Nada additions and fixes, docs and logging in general for Nillion (#24)
* added nada program for muscle perfomance

* reverted changes in env file

* changed route names and added logic to store each for thrombosis and muscle performance

* fixed snp and genotype values

* added binaries for thrombosis and muscle-performance nada programs

* reverted env file

* Added instructions for muscle performance program

* getting dealererror on invoking/computations/thrombosis

* DealerError when adding secrets from other programs

* Reverted app.py changes, corrected thrombosis code, added comments and docs.

* Mostly reverted changes to README.

---------

+++ services/nillion-interactor/programs/thrombosis.py
@@ -1,13 +1,16 @@
 from nada_dsl import *
 
-
+# Checks based on SNP rs6025 whether someone has risk for thrombosis
+# Genotype GG indicates normal risk of thrombosis
+# We arbitrarily assume elevated risk otherwise
+# This program returns 0 if normal risk, 1 if elevated risk
 def nada_main():
     party1 = Party(name="Party1")
     snp = SecretInteger(Input(name="snp", party=party1))
     genotype = SecretInteger(Input(name="genotype", party=party1))
 
-    target_snp = Integer(6)
-    target_genotype = Integer(12)
+    target_snp = Integer(12)
+    target_genotype = Integer(7)
 
     eq1 = (snp >= target_snp).if_else((target_snp >= snp).if_else(Integer(1), Integer(0)), Integer(0))
     eq2 = (genotype >= target_genotype).if_else((target_genotype >= genotype).if_else(Integer(1), Integer(0)), Integer(0))
